requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astrapy==0.7.6
cassandra-driver==3.29.0
cassio==0.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
fastapi==0.109.2
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
numpy==1.26.4
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator
passlib[bcrypt]